search-weaviate/requirements.txt
fixieai ~= 0.2.21
weaviate-client == 3.18.0